requirements.txt
dbt-core~=1.7.10
dbt-snowflake~=1.7.2
pandas~=1.5
python-dotenv~=0.21.1
snowflake-connector-python~=3.7.1
slack-sdk~=3.27.2